Nebulosa caraio! Efeito fogo arrumado, ver o bagulho da vida como float --'

+++ Space Conqueror/Assets/Scripts/Effects/PlayerEffects.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public abstract class PlayerEffects
 {
-    public int _damage;
+    public float _damage;
     public float _timer = 0f;
     public float _speed;
     public int _flameDamage = 5;
